moved addAppMessageToState to notebook-reducer per @bcolloran's request
